Reapply "first draft (needs tests)"
This reverts commit 496a9fb3b993cd35151a65a6aef57aaf1c9f432f.
